requirements.txt
Flask>=2.3
pymongo>=4.5
dnspython>=2.4
requests>=2.31
beautifulsoup4>=4.12
python-dateutil>=2.8
